# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Moldova Moto Tours API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Tour {
  id: number;
  name: string;
  slug: string;
  description: string;
  shortDescription: string;
  durationDays: number;
  priceEur: number;
  maxRiders: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  highlights?: string | null;
  /** @nullable */
  included?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateTourBody {
  name: string;
  slug: string;
  description: string;
  shortDescription: string;
  durationDays: number;
  priceEur: number;
  maxRiders: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  highlights?: string | null;
  /** @nullable */
  included?: string | null;
  isActive?: boolean;
}

export interface TourDate {
  id: number;
  tourId: number;
  startDate: string;
  endDate: string;
  availableSpots: number;
  maxRiders: number;
  isActive: boolean;
  tour?: Tour;
}

export interface CreateTourDateBody {
  tourId: number;
  startDate: string;
  endDate: string;
  availableSpots: number;
  maxRiders: number;
  isActive?: boolean;
}

export interface Motorcycle {
  id: number;
  name: string;
  brand: string;
  model: string;
  year: number;
  engineCC: number;
  description: string;
  /** @nullable */
  specs?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isAvailable: boolean;
}

export interface CreateMotorcycleBody {
  name: string;
  brand: string;
  model: string;
  year: number;
  engineCC: number;
  description: string;
  /** @nullable */
  specs?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isAvailable?: boolean;
}

export type BookingPaymentType =
  (typeof BookingPaymentType)[keyof typeof BookingPaymentType];

export const BookingPaymentType = {
  deposit: "deposit",
  full: "full",
} as const;

export type BookingPaymentStatus =
  (typeof BookingPaymentStatus)[keyof typeof BookingPaymentStatus];

export const BookingPaymentStatus = {
  pending: "pending",
  paid: "paid",
  refunded: "refunded",
  cancelled: "cancelled",
} as const;

export type BookingBookingStatus =
  (typeof BookingBookingStatus)[keyof typeof BookingBookingStatus];

export const BookingBookingStatus = {
  pending: "pending",
  approved: "approved",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  tourId: number;
  tourDateId: number;
  /** @nullable */
  motorcycleId?: number | null;
  fullName: string;
  email: string;
  phone: string;
  country: string;
  ridingExperience: string;
  hasLicense: boolean;
  paymentType: BookingPaymentType;
  paymentStatus: BookingPaymentStatus;
  bookingStatus: BookingBookingStatus;
  totalAmount: number;
  depositAmount: number;
  /** @nullable */
  stripePaymentIntentId?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  tour?: Tour;
  tourDate?: TourDate;
}

export type CreateBookingBodyPaymentType =
  (typeof CreateBookingBodyPaymentType)[keyof typeof CreateBookingBodyPaymentType];

export const CreateBookingBodyPaymentType = {
  deposit: "deposit",
  full: "full",
} as const;

export interface CreateBookingBody {
  tourId: number;
  tourDateId: number;
  /** @nullable */
  motorcycleId?: number | null;
  fullName: string;
  email: string;
  phone: string;
  country: string;
  ridingExperience: string;
  hasLicense: boolean;
  paymentType: CreateBookingBodyPaymentType;
  /** @nullable */
  notes?: string | null;
}

export type UpdateBookingStatusBodyBookingStatus =
  (typeof UpdateBookingStatusBodyBookingStatus)[keyof typeof UpdateBookingStatusBodyBookingStatus];

export const UpdateBookingStatusBodyBookingStatus = {
  pending: "pending",
  approved: "approved",
  cancelled: "cancelled",
} as const;

export type UpdateBookingStatusBodyPaymentStatus =
  (typeof UpdateBookingStatusBodyPaymentStatus)[keyof typeof UpdateBookingStatusBodyPaymentStatus];

export const UpdateBookingStatusBodyPaymentStatus = {
  pending: "pending",
  paid: "paid",
  refunded: "refunded",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingStatusBody {
  bookingStatus?: UpdateBookingStatusBodyBookingStatus;
  paymentStatus?: UpdateBookingStatusBodyPaymentStatus;
}

export type CreatePaymentIntentBodyPaymentType =
  (typeof CreatePaymentIntentBodyPaymentType)[keyof typeof CreatePaymentIntentBodyPaymentType];

export const CreatePaymentIntentBodyPaymentType = {
  deposit: "deposit",
  full: "full",
} as const;

export interface CreatePaymentIntentBody {
  bookingId: number;
  paymentType: CreatePaymentIntentBodyPaymentType;
}

export interface PaymentIntentResponse {
  clientSecret: string;
  amount: number;
  currency: string;
}

export interface ConfirmPaymentBody {
  bookingId: number;
  paymentIntentId: string;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  token: string;
  username: string;
}

export interface ContactFormBody {
  name: string;
  email: string;
  subject: string;
  message: string;
}

export type ListTourDatesParams = {
  tourId?: number;
};