# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Moldova Moto Tours API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingBookingStatus } from "./bookingBookingStatus";
import type { BookingPaymentStatus } from "./bookingPaymentStatus";
import type { BookingPaymentType } from "./bookingPaymentType";
import type { Tour } from "./tour";
import type { TourDate } from "./tourDate";

export interface Booking {
  id: number;
  tourId: number;
  tourDateId: number;
  /** @nullable */
  motorcycleId?: number | null;
  fullName: string;
  email: string;
  phone: string;
  country: string;
  ridingExperience: string;
  hasLicense: boolean;
  paymentType: BookingPaymentType;
  paymentStatus: BookingPaymentStatus;
  bookingStatus: BookingBookingStatus;
  totalAmount: number;
  depositAmount: number;
  /** @nullable */
  stripePaymentIntentId?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  tour?: Tour;
  tourDate?: TourDate;
}