# lib/api-zod/src/generated/types/confirmPaymentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Moldova Moto Tours API
 * OpenAPI spec version: 0.1.0
 */

export interface ConfirmPaymentBody {
  bookingId: number;
  paymentIntentId: string;
}