# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Moldova Moto Tours API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookingBodyPaymentType } from "./createBookingBodyPaymentType";

export interface CreateBookingBody {
  tourId: number;
  tourDateId: number;
  /** @nullable */
  motorcycleId?: number | null;
  fullName: string;
  email: string;
  phone: string;
  country: string;
  ridingExperience: string;
  hasLicense: boolean;
  paymentType: CreateBookingBodyPaymentType;
  /** @nullable */
  notes?: string | null;
}