# lib/api-zod/src/generated/types/createPaymentIntentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Moldova Moto Tours API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePaymentIntentBodyPaymentType } from "./createPaymentIntentBodyPaymentType";

export interface CreatePaymentIntentBody {
  bookingId: number;
  paymentType: CreatePaymentIntentBodyPaymentType;
}